requirements.txt
google-genai>=1.14.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
pyyaml>=6.0.2
google-cloud-aiplatform>=1.74.0
